requirements.txt
streamlit
tensorflow
google.generativeai 
Pillow
numpy
pandas
python-dotenv
